Create MerchantMenuPage
