+++ src/main/resources/templates/HomePage.html
@@ -10,8 +10,8 @@
 <body>
 <h1>Home Page</h1>
 
-<form action="#" th:action="@{/goToAddShopPage}">
-    <input type="submit" value="Go To Add Shop Page"/>
+<form action="#" th:action="@{/goToMerchantMenuPage}">
+    <input type="submit" value="Go To Merchant Menu Page"/>
 </form>
 
 <form action="#" th:action="@{/search}">
